refactor(native): move runtime world into subdomain folder
