+++ src/native/engine/runtime/world/world_abi.c
