+++ src/native/engine/runtime/world/world_metrics.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_WORLD_METRICS_H
 #define BANANA_ENGINE_RUNTIME_WORLD_METRICS_H
 
-#include "../ai/controller.h"
-#include "../world/world.h"
+#include "../../ai/controller.h"
+#include "../../world/world.h"
 
 #ifdef __cplusplus
 extern "C"
