+++ src/native/engine/runtime/world/world_telemetry.c
