+++ tests/native/runtime_physics_world_abi_test.c
@@ -1,5 +1,5 @@
 #include "runtime/physics_abi.h"
-#include "runtime/world_abi.h"
+#include "runtime/world/world_abi.h"
 
 #include <math.h>
 #include <stdio.h>
